apps/behavior/requirements.txt
blinker==1.9.0
click==8.2.1
colorama==0.4.6
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
fpdf==1.7.2
greenlet==3.2.4
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
SQLAlchemy==2.0.43
typing_extensions==4.14.1
Werkzeug==3.1.3
mcp==1.15.0
uvicorn==0.34.2
python-dateutil==2.9.0
tzdata==2025.3
requests==2.31.1
psycopg2-binary==2.9.9
python-jose[cryptography]
